[#144] refactor: Category 추상화
